Removed internal command inheritance on Keyword. Moved GetModulesFromUsingModule to parser from symbol resolver
