+++ src/System.Management.Automation/engine/parser/SymbolResolver.cs
@@ -9,7 +9,6 @@ using System.Linq;
 using System.Linq.Expressions;
 using System.Management.Automation.Runspaces;
 using System.Runtime.CompilerServices;
-using Microsoft.PowerShell.Commands;
 using System.IO;
 
 namespace System.Management.Automation.Language
@@ -272,41 +271,6 @@ namespace System.Management.Automation.Language
         internal readonly Parser _parser;
         internal readonly TypeResolutionState _typeResolutionState;
 
-        [ThreadStatic]
-        private static PowerShell t_usingStatementResolvePowerShell;
-
-        private static PowerShell UsingStatementResolvePowerShell
-        {
-            get
-            {
-                // The goal is to re-use runspaces, because creating runspace is an expensive part in creating PowerShell instance.
-                if (t_usingStatementResolvePowerShell == null)
-                {
-                    if (Runspace.DefaultRunspace != null)
-                    {
-                        t_usingStatementResolvePowerShell = PowerShell.Create(RunspaceMode.CurrentRunspace);
-                    }
-                    else
-                    {
-                        // Create empty iss and populate only commands, that we want to use.
-                        InitialSessionState iss = InitialSessionState.Create();
-                        iss.Commands.Add(new SessionStateCmdletEntry("Get-Module", typeof(GetModuleCommand), null));
-                        var sessionStateProviderEntry = new SessionStateProviderEntry(FileSystemProvider.ProviderName, typeof(FileSystemProvider), null);
-                        var snapin = PSSnapInReader.ReadEnginePSSnapIns().FirstOrDefault(snapIn => snapIn.Name.Equals("Microsoft.PowerShell.Core", StringComparison.OrdinalIgnoreCase));
-                        sessionStateProviderEntry.SetPSSnapIn(snapin);
-                        iss.Providers.Add(sessionStateProviderEntry);
-                        t_usingStatementResolvePowerShell = PowerShell.Create(iss);
-                    }
-                }
-                else if (Runspace.DefaultRunspace != null && t_usingStatementResolvePowerShell.Runspace != Runspace.DefaultRunspace)
-                {
-                    t_usingStatementResolvePowerShell = PowerShell.Create(RunspaceMode.CurrentRunspace);
-                }
-
-                return t_usingStatementResolvePowerShell;
-            }
-        }
-
         private SymbolResolver(Parser parser, TypeResolutionState typeResolutionState)
         {
             _symbolTable = new SymbolTable(parser);
@@ -421,97 +385,6 @@ namespace System.Management.Automation.Language
             return AstVisitAction.Continue;
         }
 
-        /// <summary>
-        /// Resolves using module to a collection of PSModuleInfos. Doesn't throw.
-        /// PSModuleInfo objects are returned in the right order: i.e. if multiply versions of the module
-        /// is presented on the system and user didn't specify version, we will return all of them, but newer one would go first.
-        /// </summary>
-        /// <param name="usingStatementAst">using statement</param>
-        /// <param name="exception">If exception happens, return exception object.</param>
-        /// <param name="wildcardCharactersUsed">
-        /// True if in the module name uses wildcardCharacter.
-        /// We don't want to resolve any wild-cards in using module.
-        /// </param>
-        /// <param name="isConstant">True if module hashtable contains constant value (it's our requirement).</param>
-        /// <returns>Modules, if can resolve it. null if any problems happens.</returns>
-        private Collection<PSModuleInfo> GetModulesFromUsingModule(UsingStatementAst usingStatementAst, out Exception exception, out bool wildcardCharactersUsed, out bool isConstant)
-        {
-            exception = null;
-            wildcardCharactersUsed = false;
-            isConstant = true;
-
-            // fullyQualifiedName can be string or hashtable
-            object fullyQualifiedName;
-            if (usingStatementAst.ModuleSpecification != null)
-            {
-                object resultObject;
-                if (!IsConstantValueVisitor.IsConstant(usingStatementAst.ModuleSpecification, out resultObject, forAttribute: false, forRequires: true))
-                {
-                    isConstant = false;
-                    return null;
-                }
-
-                var hashtable = resultObject as System.Collections.Hashtable;
-                var ms = new ModuleSpecification();
-                exception = ModuleSpecification.ModuleSpecificationInitHelper(ms, hashtable);
-                if (exception != null)
-                {
-                    return null;
-                }
-
-                if (WildcardPattern.ContainsWildcardCharacters(ms.Name))
-                {
-                    wildcardCharactersUsed = true;
-                    return null;
-                }
-
-                fullyQualifiedName = ms;
-            }
-            else
-            {
-                string fullyQualifiedNameStr = usingStatementAst.Name.Value;
-
-                if (WildcardPattern.ContainsWildcardCharacters(fullyQualifiedNameStr))
-                {
-                    wildcardCharactersUsed = true;
-                    return null;
-                }
-
-                // case 1: relative path. Relative for file in the same folder should include .\
-                bool isPath = fullyQualifiedNameStr.Contains(@"\");
-                if (isPath && !LocationGlobber.IsAbsolutePath(fullyQualifiedNameStr))
-                {
-                    string rootPath = Path.GetDirectoryName(_parser._fileName);
-                    if (rootPath != null)
-                    {
-                        fullyQualifiedNameStr = Path.Combine(rootPath, fullyQualifiedNameStr);
-                    }
-                }
-
-                // case 2: Module by name
-                // case 3: Absolute Path
-                // We don't need to do anything for these cases, FullyQualifiedName already handle it.
-
-                fullyQualifiedName = fullyQualifiedNameStr;
-            }
-
-            var commandInfo = new CmdletInfo("Get-Module", typeof(GetModuleCommand));
-            // TODO(sevoroby): we should consider an async call with cancellation here.
-            UsingStatementResolvePowerShell.Commands.Clear();
-            try
-            {
-                return UsingStatementResolvePowerShell.AddCommand(commandInfo)
-                    .AddParameter("FullyQualifiedName", fullyQualifiedName)
-                    .AddParameter("ListAvailable", true)
-                    .Invoke<PSModuleInfo>();
-            }
-            catch (Exception e)
-            {
-                exception = e;
-                return null;
-            }
-        }
-
         public override AstVisitAction VisitUsingStatement(UsingStatementAst usingStatementAst)
         {
             if (usingStatementAst.UsingStatementKind == UsingStatementKind.Module)
@@ -519,7 +392,7 @@ namespace System.Management.Automation.Language
                 Exception exception;
                 bool wildcardCharactersUsed;
                 bool isConstant;
-                var moduleInfo = GetModulesFromUsingModule(usingStatementAst, out exception, out wildcardCharactersUsed, out isConstant);
+                var moduleInfo = _parser.GetModulesFromUsingModule(usingStatementAst, out exception, out wildcardCharactersUsed, out isConstant);
                 if (!isConstant)
                 {
                     _parser.ReportError(usingStatementAst.Extent, () => ParserStrings.RequiresArgumentMustBeConstant);
